backend/requirements.txt
altgraph==0.17.4
asgiref==3.8.1
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
crispy-bootstrap5==2025.4
cryptography==36.0.2
distlib==0.3.9
Django==5.2
django-crispy-forms==2.4
filelock==3.18.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==24.2
pefile==2023.2.7
platformdirs==4.3.7
pycparser==2.22
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
PyJWT==2.10.1
PyMySQL==1.1.1
PySocks==1.7.1
python-dotenv==1.1.0
pywin32-ctypes==0.2.3
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.40
sqlparse==0.5.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.30.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
